refactor: widget domain으로 분리
