+++ server/src/main/java/com/seniclass/server/domain/widget/dto/AssignmentSubmissionStatsDto.java
@@ -1,4 +1,4 @@
-package com.seniclass.server.domain.student.dto;
+package com.seniclass.server.domain.widget.dto;
 
 /** 과제 제출률 통계 조회 결과 */
 public record AssignmentSubmissionStatsDto(
