+++ server/src/main/java/com/seniclass/server/domain/widget/dto/StudentVideoCountDto.java
@@ -1,4 +1,4 @@
-package com.seniclass.server.domain.student.dto;
+package com.seniclass.server.domain.widget.dto;
 
 /** 학생별 비디오 수강 개수 조회 결과 */
 public record StudentVideoCountDto(Long studentId, Long videoCount) {}
