+++ server/src/main/java/com/seniclass/server/domain/widget/service/LectureStatisticsServiceImpl.java
@@ -1,18 +1,18 @@
-package com.seniclass.server.domain.student.service;
+package com.seniclass.server.domain.widget.service;
 
 import com.seniclass.server.domain.lecture.domain.Lecture;
 import com.seniclass.server.domain.lecture.repository.LectureRepository;
 import com.seniclass.server.domain.lecture.repository.VideoRepository;
-import com.seniclass.server.domain.student.dto.AgeGroupGenderStatsDto;
-import com.seniclass.server.domain.student.dto.AssignmentSubmissionStatsDto;
-import com.seniclass.server.domain.student.dto.StudentVideoCountDto;
-import com.seniclass.server.domain.student.dto.response.LectureStatisticsResponse;
-import com.seniclass.server.domain.student.dto.response.LectureStatisticsResponse.AgeGroupStatistics;
-import com.seniclass.server.domain.student.dto.response.LectureStatisticsResponse.AssignmentSubmissionStatistics;
-import com.seniclass.server.domain.student.dto.response.LectureStatisticsResponse.AssignmentSubmissionStatistics.AssignmentSubmissionRate;
-import com.seniclass.server.domain.student.dto.response.LectureStatisticsResponse.VideoStatistics;
 import com.seniclass.server.domain.student.enums.Gender;
-import com.seniclass.server.domain.student.repository.LectureStatisticsRepository;
+import com.seniclass.server.domain.widget.dto.AgeGroupGenderStatsDto;
+import com.seniclass.server.domain.widget.dto.AssignmentSubmissionStatsDto;
+import com.seniclass.server.domain.widget.dto.StudentVideoCountDto;
+import com.seniclass.server.domain.widget.dto.response.LectureStatisticsResponse;
+import com.seniclass.server.domain.widget.dto.response.LectureStatisticsResponse.AgeGroupStatistics;
+import com.seniclass.server.domain.widget.dto.response.LectureStatisticsResponse.AssignmentSubmissionStatistics;
+import com.seniclass.server.domain.widget.dto.response.LectureStatisticsResponse.AssignmentSubmissionStatistics.AssignmentSubmissionRate;
+import com.seniclass.server.domain.widget.dto.response.LectureStatisticsResponse.VideoStatistics;
+import com.seniclass.server.domain.widget.repository.LectureStatisticsRepository;
 import com.seniclass.server.global.exception.CommonException;
 import com.seniclass.server.global.exception.errorcode.LectureErrorCode;
 import java.util.ArrayList;
